# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daybook digital planner platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sticker-packs:
    get:
      operationId: listStickerPacks
      tags: [stickerPacks]
      summary: List sticker packs
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, live, all]
      responses:
        "200":
          description: Sticker packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StickerPack"
    post:
      operationId: createStickerPack
      tags: [stickerPacks]
      summary: Create sticker pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StickerPackInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerPack"
  /sticker-packs/{id}:
    get:
      operationId: getStickerPack
      tags: [stickerPacks]
      summary: Get sticker pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sticker pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerPack"
    patch:
      operationId: updateStickerPack
      tags: [stickerPacks]
      summary: Update sticker pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StickerPackUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerPack"
    delete:
      operationId: deleteStickerPack
      tags: [stickerPacks]
      summary: Delete sticker pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /sticker-packs/{id}/publish:
    post:
      operationId: publishStickerPack
      tags: [stickerPacks]
      summary: Publish sticker pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerPack"
  /sticker-packs/{id}/unpublish:
    post:
      operationId: unpublishStickerPack
      tags: [stickerPacks]
      summary: Unpublish sticker pack
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerPack"
  /sticker-packs/ai-draft:
    post:
      operationId: aiDraftStickerPack
      tags: [stickerPacks]
      summary: AI-draft sticker pack concepts and copy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiDraftPackInput"
      responses:
        "200":
          description: AI draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiDraftPackResult"

  # ── INSERTS ──────────────────────────────────────────────────────────────────
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List related products
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, live, all]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatedProduct"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create related product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelatedProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedProduct"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get related product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedProduct"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update related product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelatedProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedProduct"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete related product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /products/{id}/publish:
    post:
      operationId: publishProduct
      tags: [products]
      summary: Publish product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedProduct"
  /products/{id}/unpublish:
    post:
      operationId: unpublishProduct
      tags: [products]
      summary: Unpublish product
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelatedProduct"

  # ── EDITIONS ─────────────────────────────────────────────────────────────────
...
  /sync/calendar/pull:
    post:
      operationId: pullCalendarEvents
      tags: [googleSync]
      summary: Pull Google Calendar events into planner day/week views
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarSyncInput"
      responses:
        "200":
          description: Pulled events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
  /sync/calendar/push:
    post:
      operationId: pushPlannerToCalendar
      tags: [googleSync]
      summary: Push planner blocks to Google Calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarPushInput"
      responses:
        "200":
          description: Push result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
  /sync/docs/push:
    post:
      operationId: pushNotesToDocs
      tags: [googleSync]
      summary: Push notes / brain-dump to Google Docs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocsPushInput"
      responses:
        "200":
          description: Push result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
...
components:
  schemas:
...
    CatalogStatus:
      type: string
      enum: [draft, live]

    # ── USER ──────────────────────────────────────────────────────────────────
...
    StickerPack:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        slug:           { type: string }
        description:    { type: ["string", "null"] }
        status:         { $ref: "#/components/schemas/CatalogStatus" }
        category:       { type: ["string", "null"] }
        previewImageUrl: { type: ["string", "null"] }
        stickerCount:   { type: ["integer", "null"] }
        driveFileId:    { type: ["string", "null"] }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    StickerPackInput:
      type: object
      required: [name]
      properties:
        name:           { type: string, minLength: 1 }
        slug:           { type: string }
        description:    { type: string }
        category:       { type: string }
        previewImageUrl: { type: string }
        stickerCount:   { type: integer }
    StickerPackUpdate:
      type: object
      properties:
        name:           { type: string, minLength: 1 }
        description:    { type: string }
        category:       { type: string }
        previewImageUrl: { type: string }
        stickerCount:   { type: integer }
        driveFileId:    { type: string }
    AiDraftPackInput:
      type: object
      required: [concept]
      properties:
        concept:  { type: string }
        style:    { type: string }
        audience: { type: string }
    AiDraftPackResult:
      type: object
      required: [name, description, suggestedCount, concepts]
      properties:
        name:           { type: string }
        description:    { type: string }
        suggestedCount: { type: integer }
        concepts:       { type: array, items: { type: string } }

    # ── INSERT ────────────────────────────────────────────────────────────────
...
    RelatedProduct:
      type: object
      required: [id, name, slug, type, status, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        slug:           { type: string }
        description:    { type: ["string", "null"] }
        status:         { $ref: "#/components/schemas/CatalogStatus" }
        type:
          type: string
          enum: [notes-only, to-do, tracker, mixed]
        previewImageUrl: { type: ["string", "null"] }
        price:          { type: ["number", "null"] }
        driveFileId:    { type: ["string", "null"] }
        createdAt:      { type: string, format: date-time }
        updatedAt:      { type: string, format: date-time }
    RelatedProductInput:
      type: object
      required: [name, type]
      properties:
        name:           { type: string, minLength: 1 }
        slug:           { type: string }
        description:    { type: string }
        type:
          type: string
          enum: [notes-only, to-do, tracker, mixed]
        previewImageUrl: { type: string }
        price:          { type: number }
    RelatedProductUpdate:
      type: object
      properties:
        name:           { type: string, minLength: 1 }
        description:    { type: string }
        type:
          type: string
          enum: [notes-only, to-do, tracker, mixed]
        previewImageUrl: { type: string }
        price:          { type: number }
        driveFileId:    { type: string }

    # ── EDITION ───────────────────────────────────────────────────────────────
...
    CalendarSyncInput:
      type: object
      required: [plannerConfigId]
      properties:
        plannerConfigId: { type: integer }
        dateFrom:        { type: string, format: date }
        dateTo:          { type: string, format: date }
    CalendarPushInput:
      type: object
      required: [plannerConfigId]
      properties:
        plannerConfigId: { type: integer }
        blocks:
          type: array
          items:
            type: object
            required: [title, startDate, endDate]
            properties:
              title:     { type: string }
              startDate: { type: string, format: date }
              endDate:   { type: string, format: date }
    DocsPushInput:
      type: object
      required: [title, content]
      properties:
        title:   { type: string }
        content: { type: string }
    SyncResult:
      type: object
      required: [success, syncedAt]
      properties:
        success:   { type: boolean }
        syncedAt:  { type: string, format: date-time }
        itemCount: { type: ["integer", "null"] }
        message:   { type: ["string", "null"] }

    # ── AI PROXY ─────────────────────────────────────────────────────────────
...